refactor: rename module `_future_cdk` to `shared`
