+++ airbyte/_processors/sql/postgres.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 
 from overrides import overrides
 
-from airbyte._future_cdk.sql_processor import SqlConfig, SqlProcessorBase
 from airbyte._writers.jsonl import JsonlWriter
 from airbyte.secrets.base import SecretString
+from airbyte.shared.sql_processor import SqlConfig, SqlProcessorBase
 
 
 class PostgresConfig(SqlConfig):
