+++ airbyte/_writers/file_writers.py
@@ -22,10 +22,10 @@ if TYPE_CHECKING:
         AirbyteRecordMessage,
     )
 
-    from airbyte._future_cdk.catalog_providers import CatalogProvider
-    from airbyte._future_cdk.state_writers import StateWriterBase
     from airbyte._message_iterators import AirbyteMessageIterator
     from airbyte.progress import ProgressTracker
+    from airbyte.shared.catalog_providers import CatalogProvider
+    from airbyte.shared.state_writers import StateWriterBase
     from airbyte.strategies import WriteStrategy
