+++ airbyte/shared/catalog_providers.py
