+++ airbyte/shared/sql_processor.py
@@ -41,7 +41,6 @@ from airbyte_protocol.models import (
 )
 
 from airbyte import exceptions as exc
-from airbyte._future_cdk.state_writers import StdOutStateWriter
 from airbyte._util.name_normalizers import LowerCaseNormalizer
 from airbyte.constants import (
     AB_EXTRACTED_AT_COLUMN,
@@ -50,6 +49,7 @@ from airbyte.constants import (
     DEBUG_MODE,
 )
 from airbyte.records import StreamRecordHandler
+from airbyte.shared.state_writers import StdOutStateWriter
 from airbyte.strategies import WriteMethod, WriteStrategy
 from airbyte.types import SQLTypeConverter
 
@@ -64,11 +64,11 @@ if TYPE_CHECKING:
     from sqlalchemy.sql.type_api import TypeEngine
 
     from airbyte._batch_handles import BatchHandle
-    from airbyte._future_cdk.catalog_providers import CatalogProvider
-    from airbyte._future_cdk.state_writers import StateWriterBase
     from airbyte._writers.jsonl import FileWriterBase
     from airbyte.progress import ProgressTracker
     from airbyte.secrets.base import SecretString
+    from airbyte.shared.catalog_providers import CatalogProvider
+    from airbyte.shared.state_writers import StateWriterBase
 
 
 class RecordDedupeMode(enum.Enum):
